Rename and move album classes to albums module
