+++ src/app/components/albums/album.rs
