+++ src/app/components/albums/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: © 2024 David Bliss
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+pub mod album;
+pub mod album_filter;
+pub mod folders_album;
+pub mod months_album;
+pub mod years_album;
