+++ src/app/components/mod.rs
@@ -4,13 +4,9 @@
 
 pub mod about;
 pub mod preferences;
-pub mod album;
-pub mod album_filter;
-pub mod month_photos;
-pub mod year_photos;
+pub mod albums;
 pub mod one_photo;
 pub mod photo_info;
-pub mod folder_photos;
 pub mod library;
 pub mod progress_monitor;
 pub mod progress_panel;
